+++ lib/segment/segment.types.ts
@@ -1,0 +1,3 @@
+//TO-DO: Detail type SegmentObject
+
+export type SegmentObject = any;